+++ src/components/Charts/BarChart/BarChartContent.tsx
@@ -9,7 +9,6 @@ import ActivityIndicator from '@components/ActivityIndicator';
 import ChartTooltipLayer from '@components/Charts/components/ChartTooltipLayer';
 import ChartXAxisLabels from '@components/Charts/components/ChartXAxisLabels';
 import ChartYAxisLabels from '@components/Charts/components/ChartYAxisLabels';
-import {CHART_CONTENT_MIN_HEIGHT, GLYPH_PADDING} from '@components/Charts/constants';
 import type {HitTestArgs} from '@components/Charts/hooks';
 import {
     useChartFontManager,
@@ -22,7 +21,7 @@ import {
     useYAxisLabelWidth,
 } from '@components/Charts/hooks';
 import {calculateMinDomainPadding} from '@components/Charts/utils';
-import VictoryTheme from '@components/Charts/VictoryTheme';
+import VictoryTheme, {CHART_CONTENT_MIN_HEIGHT, GLYPH_PADDING} from '@components/Charts/VictoryTheme';
 import useTheme from '@hooks/useTheme';
 import useThemeStyles from '@hooks/useThemeStyles';
 import type {SkeletonSpanReasonAttributes} from '@libs/telemetry/useSkeletonSpan';
