+++ src/components/Charts/components/ChartYAxisLabels.tsx
@@ -2,10 +2,9 @@ import {Paragraph} from '@shopify/react-native-skia';
 import type {SkTypefaceFontProvider} from '@shopify/react-native-skia';
 import React from 'react';
 import type {ChartBounds, Scale} from 'victory-native';
-import {GLYPH_PADDING, MAX_Y_AXIS_LABEL_WIDTH} from '@components/Charts/constants';
 import {useChartParagraphs} from '@components/Charts/hooks';
 import {getFontLineMetrics} from '@components/Charts/utils';
-import VictoryTheme from '@components/Charts/VictoryTheme';
+import VictoryTheme, {GLYPH_PADDING, MAX_Y_AXIS_LABEL_WIDTH} from '@components/Charts/VictoryTheme';
 
 type ChartYAxisLabelsProps = {
     /** Tick values on the Y axis. */
