+++ tests/unit/components/Charts/useChartLabelLayout.test.ts
@@ -1,10 +1,10 @@
 import type {SkTypefaceFontProvider} from '@shopify/react-native-skia';
 import {renderHook} from '@testing-library/react-native';
-import {SIN_45} from '@components/Charts/constants';
 import useChartLabelLayout from '@components/Charts/hooks/useChartLabelLayout';
 import useChartLabelMeasurements from '@components/Charts/hooks/useChartLabelMeasurements';
 import type {ChartDataPoint} from '@components/Charts/types';
 import type * as ChartUtils from '@components/Charts/utils';
+import {SIN_45} from '@components/Charts/VictoryTheme';
 
 /**
  * Each glyph = PX_PER_CHAR wide. This gives deterministic widths:
